Updated section-looks-overview.liquid
Had to change the logic behind looks rendering. Previously the model object was associated with a look object, but due to shopify loop limits that is a bad proccedure, which is why I added a new list metafield to the model object, where looks will now be added aswell. Code was adjusted accordingly on both sections

Also small bugfix + CSS fixes